Moved example application under tests
