+++ flink-connector-prometheus/src/test/java/org/apache/flink/connector/prometheus/sink/examples/DataStreamExample.java
@@ -15,15 +15,17 @@
  *  limitations under the License.
  */
 
-package org.apache.flink.connector.prometheus.examples;
+package org.apache.flink.connector.prometheus.sink.examples;
 
+import org.apache.flink.api.common.typeinfo.TypeInformation;
+import org.apache.flink.api.java.typeutils.GenericTypeInfo;
+import org.apache.flink.api.java.typeutils.ResultTypeQueryable;
 import org.apache.flink.api.java.utils.ParameterTool;
 import org.apache.flink.connector.base.sink.AsyncSinkBase;
 import org.apache.flink.connector.prometheus.sink.PrometheusRequestSigner;
 import org.apache.flink.connector.prometheus.sink.PrometheusSink;
 import org.apache.flink.connector.prometheus.sink.PrometheusTimeSeries;
 import org.apache.flink.connector.prometheus.sink.PrometheusTimeSeriesLabelsAndMetricNameKeySelector;
-import org.apache.flink.connector.prometheus.sink.aws.AmazonManagedPrometheusWriteRequestSigner;
 import org.apache.flink.connector.prometheus.sink.errorhandling.OnErrorBehavior;
 import org.apache.flink.connector.prometheus.sink.errorhandling.SinkWriterErrorHandlingBehaviorConfiguration;
 import org.apache.flink.connector.prometheus.sink.http.RetryConfiguration;
@@ -31,15 +33,28 @@ import org.apache.flink.connector.prometheus.sink.prometheus.Types;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.streaming.api.functions.source.SourceFunction;
+import org.apache.flink.util.Preconditions;
 
+import org.apache.commons.lang3.RandomUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.io.Serializable;
 import java.util.function.Supplier;
 
-/** Test application testing the Prometheus sink connector. */
-public class DataStreamJob {
-    private static final Logger LOGGER = LoggerFactory.getLogger(DataStreamJob.class);
+/**
+ * Example application demonstrating the usage of the Prometheus sink connector.
+ *
+ * <p>The application expects a single configuration parameter, with the RemoteWrite endpoint URL:
+ * --prometheusRemoteWriteUrl &lt;URL&gt;
+ *
+ * <p>The application generates random TimeSeries internally and sinks to Prometheus.
+ *
+ * <p>It also demonstrates how to keyBy the stream before the sink, to support a sink parallelism
+ * greater than 1.
+ */
+public class DataStreamExample {
+    private static final Logger LOGGER = LoggerFactory.getLogger(DataStreamExample.class);
 
     public static void main(String[] args) throws Exception {
         StreamExecutionEnvironment env = StreamExecutionEnvironment.getExecutionEnvironment();
@@ -55,16 +70,16 @@ public class DataStreamJob {
         String prometheusRemoteWriteUrl = applicationParameters.get("prometheusRemoteWriteUrl");
         LOGGER.info("Prometheus URL:{}", prometheusRemoteWriteUrl);
 
-        // Optionally configure Amazon Managed Prometheus request signer
+        // Optionally configure a request signer
         PrometheusRequestSigner requestSigner = null;
-        String ampAWSRegion = applicationParameters.get("awsRegion");
-        if (ampAWSRegion != null) {
-            requestSigner =
-                    new AmazonManagedPrometheusWriteRequestSigner(
-                            prometheusRemoteWriteUrl, ampAWSRegion);
-            LOGGER.info(
-                    "Enable Amazon Managed Prometheus request-signer, region: {}", ampAWSRegion);
-        }
+
+        // Uncomment the following line to enable the AmazonManagedPrometheusWriteRequestSigner
+        // for signing requests when writing to Amazon Managed Prometheus
+        //        requestSigner =
+        //                new AmazonManagedPrometheusWriteRequestSigner(
+        //                        prometheusRemoteWriteUrl, applicationParameters.get("awsRegion"));
+        //                    new AmazonManagedPrometheusWriteRequestSigner(
+        //                            prometheusRemoteWriteUrl, ampAWSRegion);
 
         // Configure data generator
         int generatorMinSamplesPerTimeSeries = 1;
@@ -132,10 +147,114 @@ public class DataStreamJob {
                         .setMetricGroupName("Prometheus") // Optional, default "Prometheus"
                         .build();
 
+        // Key the stream using the provided KeySelector, before sinking to Prometheus.
+        // This is required when running with parallelism > 1, to guarantee that all the TimeSeries
+        // with the same set of
+        // Labels are written by the same subtask, to prevent accidental out-of-order writes.
         prometheusTimeSeries
                 .keyBy(new PrometheusTimeSeriesLabelsAndMetricNameKeySelector())
                 .sinkTo(sink);
 
         env.execute("Prometheus Sink test");
+    }
+
+    /**
+     * Simple data generator. Generates records continuously, with a fixed delay, using a record
+     * Supplier.
+     */
+    public static class FixedDelayDataGeneratorSource<T>
+            implements SourceFunction<T>, ResultTypeQueryable<T> {
+        private volatile boolean isRunning = true;
+
+        private final Supplier<T> eventGenerator;
+        private final long pauseMillis;
+        private final Class<T> payloadClass;
+
+        public FixedDelayDataGeneratorSource(
+                Class<T> payloadClass, Supplier<T> eventGenerator, long pauseMillis) {
+            Preconditions.checkArgument(
+                    pauseMillis > 0,
+                    "Pause between time-series must be > 0"); // If zero, the source may generate
+            // duplicate timestamps
+            this.eventGenerator = eventGenerator;
+            this.pauseMillis = pauseMillis;
+            this.payloadClass = payloadClass;
+        }
+
+        @Override
+        public void run(SourceContext<T> sourceContext) throws Exception {
+            while (isRunning) {
+                T event = eventGenerator.get();
+                sourceContext.collect(event);
+                Thread.sleep(pauseMillis);
+            }
+        }
+
+        @Override
+        public void cancel() {
+            isRunning = false;
+        }
+
+        @Override
+        public TypeInformation<T> getProducedType() {
+            return new GenericTypeInfo<>(payloadClass);
+        }
+    }
+
+    /**
+     * Simple dummy TimeSeries generator, generating random samples for a given number of "metrics"
+     * and a given number of "sources".
+     *
+     * <p>Sample timestamp is always the current system time. The value is random, between 0 and 1.
+     */
+    public static class SimpleRandomMetricTimeSeriesGenerator implements Serializable {
+
+        private final int numberOfSources;
+        private final int minNrOfSamples;
+        private final int maxNrOfSamples;
+        private final short numberOfMetricsPerSource;
+
+        public SimpleRandomMetricTimeSeriesGenerator(
+                int minNrOfSamples,
+                int maxNrOfSamples,
+                int numberOfSources,
+                short numberOfMetricsPerSource) {
+            this.numberOfSources = numberOfSources;
+            this.minNrOfSamples = minNrOfSamples;
+            this.maxNrOfSamples = maxNrOfSamples;
+            this.numberOfMetricsPerSource = numberOfMetricsPerSource;
+        }
+
+        private String randomMetricName() {
+            short metricNr = (short) (RandomUtils.nextDouble(0, 1) * numberOfMetricsPerSource);
+            return String.format("M%05d", metricNr);
+        }
+
+        private String randomSourceId() {
+            int sourceId = RandomUtils.nextInt(0, numberOfSources);
+            return String.format("S%010d", sourceId);
+        }
+
+        private PrometheusTimeSeries nextTimeSeries(int minNrOfSamples, int maxNrOfSamples) {
+            String instanceId = randomSourceId();
+
+            int nrOfSamples = RandomUtils.nextInt(minNrOfSamples, maxNrOfSamples + 1);
+            String metricName = randomMetricName();
+
+            PrometheusTimeSeries.Builder builder =
+                    PrometheusTimeSeries.builder()
+                            .withMetricName(metricName)
+                            .addLabel("SourceID", instanceId);
+            for (int i = 0; i < nrOfSamples; i++) {
+                double value = RandomUtils.nextDouble(0, 1);
+                builder.addSample(value, System.currentTimeMillis());
+            }
+            return builder.build();
+        }
+
+        public Supplier<PrometheusTimeSeries> generator() {
+            return (Supplier<PrometheusTimeSeries> & Serializable)
+                    () -> nextTimeSeries(minNrOfSamples, maxNrOfSamples);
+        }
     }
 }
