+++ flink-connector-prometheus/src/test/resources/log4j2.properties
